Move non-core dependencies to extensions
