+++ Project.toml
@@ -6,30 +6,30 @@ version = "0.4.2"
 [deps]
 ComponentArrays = "b0b7db55-cfe3-40fc-9ded-d10e2dbeff66"
 CxxWrap = "1f15a43c-97ca-5a2a-ae31-89f07a497df4"
-Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
-DistributionsHEP = "0d0d6f60-fb7b-48ca-90bb-e14e8b1655a7"
-FHist = "68837c9b-b678-4cd5-9925-8a54edc8f695"
 IterTools = "c8e1da08-722c-5040-9ed9-7db0dc04731e"
 Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Minuit2_Julia_Wrapper_jll = "65cff058-67fb-5034-9600-a8f6a0feb90b"
-Polynomials = "f27b6e38-b328-58d1-80ce-0feddd5e7a45"
 PrettyTables = "08abe8d2-0d0c-5749-adfa-8a2ac140af0d"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
-SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
-StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 
 [weakdeps]
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+DistributionsHEP = "0d0d6f60-fb7b-48ca-90bb-e14e8b1655a7"
+FHist = "68837c9b-b678-4cd5-9925-8a54edc8f695"
 FiniteDiff = "6a86dc24-6348-571c-b903-95158fe2bd41"
 Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 QuadGK = "1fd47b50-473d-5c70-9696-f719f8f3bcdc"
+RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
+StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 
 [extensions]
 Minuit2FHistExt = ["FHist"]
-Minuit2PlotsExt = ["Plots", "FHist"]
 Minuit2OptimizationExt = ["Optimization"]
+Minuit2PlotsExt = ["Plots", "FHist"]
+Minuit2RooFitExt = ["Distributions", "DistributionsHEP", "FHist", "StatsBase"]
+Minuit2RooFitPlotsExt = ["Distributions", "DistributionsHEP", "FHist", "Plots", "RecipesBase", "StatsBase"]
 
 [compat]
 ComponentArrays = "0.15.28"
@@ -42,17 +42,23 @@ LinearAlgebra = "1"
 Minuit2_Julia_Wrapper_jll = "0.4.0"
 Optimization = "4.4.0"
 Plots = "1"
-Polynomials = "4.0.19"
 PrettyTables = "2.4.0"
 Random = "1"
 RecipesBase = "^1"
-SpecialFunctions = "2.5.0"
 StatsBase = "0.34.4"
 julia = "1.10"
 
 [extras]
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+DistributionsHEP = "0d0d6f60-fb7b-48ca-90bb-e14e8b1655a7"
+FHist = "68837c9b-b678-4cd5-9925-8a54edc8f695"
+FiniteDiff = "6a86dc24-6348-571c-b903-95158fe2bd41"
 Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+QuadGK = "1fd47b50-473d-5c70-9696-f719f8f3bcdc"
+RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
+StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "FiniteDiff", "Optimization", "QuadGK"]
+test = ["Test", "FiniteDiff", "Optimization", "QuadGK", "Distributions", "DistributionsHEP", "FHist", "Plots", "RecipesBase", "StatsBase"]
